refacto test Database for test
